docker-requirements.txt
aioboto3==11.1.0
cryptography==41.0.3
email-validator==2.0.0
fastapi==0.100.1
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-swagger-ui==4.11.1
gunicorn==23.0.0
passlib==1.7.4
psycopg2-binary==2.9.7
pyjwt==2.8.0
pyotp==2.9.0
pytest==7.4.0
pytest-asyncio==0.21.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
requests==2.31.0
uvicorn==0.23.2
werkzeug==2.3.7
pydantic_settings==2.9.1
asyncpg==0.30.0